+++ src/SqlArtisan/Internal/Extensions/DoubleExtensions.cs
@@ -1,0 +1,9 @@
+﻿using System.Globalization;
+
+namespace SqlArtisan.Internal;
+
+internal static class DoubleExtensions
+{
+    internal static string ToInvariantString(this double value) =>
+        value.ToString(CultureInfo.InvariantCulture);
+}